refactor(embeddings): extract embeddings-search module
